+++ src/headless/headless.js
@@ -6,7 +6,6 @@ import "./plugins/adhoc.js";            // XEP-0050 Ad Hoc Commands
 import "./plugins/bookmarks/index.js";  // XEP-0199 XMPP Ping
 import "./plugins/bosh.js";             // XEP-0206 BOSH
 import "./plugins/caps/index.js";       // XEP-0115 Entity Capabilities
-import "./plugins/carbons.js";          // XEP-0280 Message Carbons
 import "./plugins/chat/index.js";       // RFC-6121 Instant messaging
 import "./plugins/chatboxes/index.js";
 import "./plugins/disco/index.js";      // XEP-0030 Service discovery